requirements.txt
afinn==0.1
tweepy==4.10.0
pymongo==4.1.1
kafka-python==2.0.2
pandas==1.4.2
dash==2.4.1
plotly==5.8.0